# graph: merge set.Ints and set.Int64s into one generic set