--- a/graph/community/louvain_directed.go
+++ b/graph/community/louvain_directed.go
@@ -593,7 +593,7 @@ func (l *directedLocalMover) deltaQ(n graph.Node) (deltaQ float64, dst int, src 
 	gamma := l.resolution
 
 	// Find communities connected to n.
-	connected := make(set.Ints)
+	connected := make(set.Ints[int])
 	// The following for loop is equivalent to:
 	//
 	//  for _, v := range l.g.From(n) {