--- a/graph/community/louvain_directed_multiplex.go
+++ b/graph/community/louvain_directed_multiplex.go
@@ -133,18 +133,18 @@ func NewDirectedLayers(layers ...graph.Directed) (DirectedLayers, error) {
 	if len(layers) == 0 {
 		return nil, nil
 	}
-	base := make(set.Int64s)
+	base := make(set.Ints[int64])
 	nodes := layers[0].Nodes()
 	for nodes.Next() {
 		base.Add(nodes.Node().ID())
 	}
 	for i, l := range layers[1:] {
-		next := make(set.Int64s)
+		next := make(set.Ints[int64])
 		nodes := l.Nodes()
 		for nodes.Next() {
 			next.Add(nodes.Node().ID())
 		}
-		if !set.Int64sEqual(base, next) {
+		if !set.IntsEqual(base, next) {
 			return nil, fmt.Errorf("community: layer ID mismatch between layers: %d", i+1)
 		}
 	}
@@ -796,7 +796,7 @@ func (l *directedMultiplexLocalMover) deltaQ(n graph.Node) (deltaQ float64, dst 
 		iterator = &dense{n: len(l.communities)}
 	} else {
 		// Find communities connected to n.
-		connected := make(set.Ints)
+		connected := make(set.Ints[int])
 		// The following for loop is equivalent to:
 		//
 		//  for i := 0; i < l.g.Depth(); i++ {