--- a/graph/community/louvain_undirected_multiplex_test.go
+++ b/graph/community/louvain_undirected_multiplex_test.go
@@ -5,6 +5,7 @@
 package community
 
 import (
+	"fmt"
 	"math"
 	"reflect"
 	"sort"
@@ -690,4 +691,31 @@ func undirectedMultiplexFrom(raw []layer) (UndirectedLayers, []float64, error) {
 		return nil, nil, err
 	}
 	return g, weights, nil
+}
+
+func BenchmarkNewUndirectedLayers(b *testing.B) {
+	for _, graphSize := range []int{1e0, 1e1, 1e3, 1e5} {
+		for _, numGraphs := range []int{1e0, 1e1} {
+			b.Run(
+				fmt.Sprintf("graphSize=%d,numGraphs=%d", graphSize, numGraphs),
+				func(b *testing.B) {
+					g := simple.NewUndirectedGraph()
+					for i := 0; i < graphSize; i++ {
+						g.AddNode(g.NewNode())
+					}
+					gs := make([]graph.Undirected, numGraphs)
+					for i := 0; i < numGraphs; i++ {
+						gs[i] = g
+					}
+
+					b.ResetTimer()
+					for i := 0; i < b.N; i++ {
+						_, err := NewUndirectedLayers(gs...)
+						if err != nil {
+							b.Fatalf("NewUndirectedLayers failed: %v", err)
+						}
+					}
+				})
+		}
+	}
 }