--- a/graph/encoding/dot/decode.go
+++ b/graph/encoding/dot/decode.go
@@ -358,7 +358,7 @@ func (gen *generator) popSubgraph() []graph.Node {
 // unique returns the set of unique nodes contained within ns.
 func unique(ns []graph.Node) []graph.Node {
 	var nodes []graph.Node
-	seen := make(set.Int64s)
+	seen := make(set.Ints[int64])
 	for _, n := range ns {
 		id := n.ID()
 		if seen.Has(id) {