@jbduncan 2024-03-11
I'll double check that all affected code has benchmarks and write any new benchmarks as needed, as discussed in https://github.com/gonum/gonum/pull/1956#issuecomment-1988326786.