@jbduncan 2024-03-16
There's a few benchmarks to add (which I'll add TODO comments for in a TODO commit to keep track of), and once they're added, I'll run all `graph` benchmarks to see the general impact of generifying the sets.