@jbduncan 2024-03-18
Here are the latest benchmarks for `graph`.

They seem to be a bit all over the place, but in summary the geomeans seem to have gone up by 0-1%.

@kortschak Let me know what you think.

<details>
  <summary>Benchmarks</summary>

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/graph/coloring
                                                       │ benchmarks_graph_old.txt │       benchmarks_graph_new.txt        │
                                                       │          sec/op          │    sec/op      vs base                │
Coloring/empty/Dsatur-8                                             3.134n ±   0%    3.159n ±  1%   +0.80% (p=0.000 n=10)
Coloring/empty/DsaturExact-8                                        3.132n ±   1%    3.175n ±  1%   +1.34% (p=0.004 n=10)
Coloring/empty/Randomized-8                                         3.142n ±   1%    3.170n ±  0%   +0.89% (p=0.001 n=10)
Coloring/empty/RecursiveLargestFirst-8                              3.147n ±   0%    3.172n ±  0%   +0.78% (p=0.000 n=10)
Coloring/empty/SanSegundo-8                                         3.145n ±   0%    3.175n ±  0%   +0.95% (p=0.000 n=10)
Coloring/empty/WelshPowell-8                                        3.151n ±   0%    3.174n ±  0%   +0.71% (p=0.001 n=10)
Coloring/singleton/Dsatur-8                                         407.3n ±   0%    410.0n ±  0%   +0.68% (p=0.001 n=10)
Coloring/singleton/DsaturExact-8                                    942.8n ±   0%    945.8n ±  0%   +0.32% (p=0.003 n=10)
Coloring/singleton/Randomized-8                                     232.3n ±   0%    233.6n ±  0%   +0.60% (p=0.000 n=10)
Coloring/singleton/RecursiveLargestFirst-8                          293.8n ±   1%    293.1n ±  0%   -0.22% (p=0.005 n=10)
Coloring/singleton/SanSegundo-8                                     407.2n ±   0%    409.4n ±  0%   +0.53% (p=0.000 n=10)
Coloring/singleton/WelshPowell-8                                    284.3n ±   0%    285.9n ±  0%   +0.56% (p=0.000 n=10)
Coloring/kite/Dsatur-8                                              3.269µ ±   0%    3.275µ ±  0%   +0.18% (p=0.005 n=10)
Coloring/kite/DsaturExact-8                                         17.61µ ±   0%    17.66µ ±  0%        ~ (p=0.063 n=10)
Coloring/kite/Randomized-8                                          1.157µ ±   0%    1.157µ ±  1%        ~ (p=0.154 n=10)
Coloring/kite/RecursiveLargestFirst-8                               2.217µ ±   0%    2.221µ ±  0%   +0.20% (p=0.023 n=10)
Coloring/kite/SanSegundo-8                                          3.967µ ±   0%    3.981µ ±  0%   +0.37% (p=0.005 n=10)
Coloring/kite/WelshPowell-8                                         1.387µ ±   0%    1.390µ ±  0%   +0.25% (p=0.005 n=10)
Coloring/triangle+1/Dsatur-8                                        2.070µ ±   0%    2.071µ ±  0%        ~ (p=0.358 n=10)
Coloring/triangle+1/DsaturExact-8                                   9.930µ ±   0%    9.967µ ±  0%   +0.37% (p=0.000 n=10)
Coloring/triangle+1/Randomized-8                                    903.1n ±  67%    807.2n ±  1%  -10.61% (p=0.014 n=10)
Coloring/triangle+1/RecursiveLargestFirst-8                         1.525µ ± 114%    1.418µ ±  0%   -7.02% (p=0.000 n=10)
Coloring/triangle+1/SanSegundo-8                                    2.412µ ±   1%    2.409µ ±  0%        ~ (p=0.641 n=10)
Coloring/triangle+1/WelshPowell-8                                   966.6n ±   0%    972.4n ±  0%   +0.59% (p=0.000 n=10)
Coloring/bipartite_halves/Dsatur-8                                  7.911µ ±   0%    7.948µ ±  0%   +0.47% (p=0.000 n=10)
Coloring/bipartite_halves/DsaturExact-8                             32.67µ ±   0%    32.75µ ±  0%        ~ (p=0.066 n=10)
Coloring/bipartite_halves/Randomized-8                              2.162µ ±   0%    2.162µ ±  0%        ~ (p=0.718 n=10)
Coloring/bipartite_halves/RecursiveLargestFirst-8                   3.957µ ±   0%    3.962µ ±  0%        ~ (p=0.077 n=10)
Coloring/bipartite_halves/SanSegundo-8                              15.50µ ±   3%    15.01µ ±  0%   -3.17% (p=0.003 n=10)
Coloring/bipartite_halves/WelshPowell-8                             2.435µ ±   2%    2.379µ ±  1%   -2.28% (p=0.001 n=10)
Coloring/bipartite_alternating/Dsatur-8                             7.828µ ±   2%    7.834µ ±  0%        ~ (p=0.424 n=10)
Coloring/bipartite_alternating/DsaturExact-8                        32.60µ ±   1%    32.64µ ±  0%        ~ (p=0.315 n=10)
Coloring/bipartite_alternating/Randomized-8                         2.143µ ±   0%    2.145µ ±  0%        ~ (p=0.157 n=10)
Coloring/bipartite_alternating/RecursiveLargestFirst-8              3.919µ ±   0%    3.928µ ±  1%   +0.23% (p=0.006 n=10)
Coloring/bipartite_alternating/SanSegundo-8                         15.12µ ±   0%    15.09µ ±  0%   -0.15% (p=0.037 n=10)
Coloring/bipartite_alternating/WelshPowell-8                        2.471µ ±   0%    2.466µ ±  0%        ~ (p=0.638 n=10)
Coloring/3/4_bipartite/Dsatur-8                                     6.362µ ±   0%    6.371µ ±  0%        ~ (p=0.254 n=10)
Coloring/3/4_bipartite/DsaturExact-8                                30.01µ ±   0%    30.04µ ±  0%        ~ (p=0.210 n=10)
Coloring/3/4_bipartite/Randomized-8                                 1.844µ ±   0%    1.843µ ±  0%        ~ (p=0.654 n=10)
Coloring/3/4_bipartite/RecursiveLargestFirst-8                      3.505µ ±   0%    3.507µ ±  0%        ~ (p=0.138 n=10)
Coloring/3/4_bipartite/SanSegundo-8                                 13.22µ ±   1%    13.26µ ±  1%        ~ (p=0.739 n=10)
Coloring/3/4_bipartite/WelshPowell-8                                2.220µ ±   0%    2.232µ ±  0%   +0.56% (p=0.000 n=10)
Coloring/cubical/Dsatur-8                                           7.957µ ±   0%    8.015µ ±  0%   +0.72% (p=0.000 n=10)
Coloring/cubical/DsaturExact-8                                      33.04µ ±   0%    33.08µ ±  0%        ~ (p=0.060 n=10)
Coloring/cubical/Randomized-8                                       2.175µ ±   1%    2.178µ ±  0%        ~ (p=0.090 n=10)
Coloring/cubical/RecursiveLargestFirst-8                            4.037µ ±   0%    4.035µ ±  0%        ~ (p=0.494 n=10)
Coloring/cubical/SanSegundo-8                                       15.20µ ±   0%    15.17µ ±  0%        ~ (p=0.072 n=10)
Coloring/cubical/WelshPowell-8                                      2.498µ ±   0%    2.495µ ±  1%        ~ (p=0.445 n=10)
Coloring/HoG33106/Dsatur-8                                          18.94µ ±   0%    18.96µ ±  0%        ~ (p=0.078 n=10)
Coloring/HoG33106/DsaturExact-8                                     72.85µ ±   0%    72.68µ ±  0%        ~ (p=0.075 n=10)
Coloring/HoG33106/Randomized-8                                      3.999µ ±   0%    4.001µ ±  0%        ~ (p=0.158 n=10)
Coloring/HoG33106/RecursiveLargestFirst-8                           8.790µ ±   1%    8.767µ ±  0%   -0.26% (p=0.002 n=10)
Coloring/HoG33106/SanSegundo-8                                      68.30µ ±   0%    67.94µ ±  0%   -0.54% (p=0.000 n=10)
Coloring/HoG33106/WelshPowell-8                                     4.975µ ±   1%    4.964µ ±  0%        ~ (p=0.143 n=10)
Coloring/HoG41237/Dsatur-8                                          214.5m ±   0%    214.1m ±  1%        ~ (p=0.247 n=10)
Coloring/HoG41237/DsaturExact-8                                     234.5m ±   1%    238.1m ±  1%   +1.54% (p=0.000 n=10)
Coloring/HoG41237/Randomized-8                                      2.403m ±   0%    2.422m ±  1%   +0.76% (p=0.000 n=10)
Coloring/HoG41237/RecursiveLargestFirst-8                           8.914m ±   1%    8.934m ±  1%        ~ (p=0.143 n=10)
Coloring/HoG41237/SanSegundo-8                                      504.6m ±   1%    503.0m ±  0%   -0.31% (p=0.019 n=10)
Coloring/HoG41237/WelshPowell-8                                     3.652m ±   1%    3.609m ±  0%   -1.18% (p=0.000 n=10)
Coloring/exams/Dsatur-8                                             4.636µ ±   1%    4.637µ ±  0%        ~ (p=0.868 n=10)
Coloring/exams/DsaturExact-8                                        26.41µ ±   0%    26.41µ ±  0%        ~ (p=0.870 n=10)
Coloring/exams/Randomized-8                                         1.508µ ±   0%    1.509µ ±  0%        ~ (p=0.187 n=10)
Coloring/exams/RecursiveLargestFirst-8                              3.044µ ±   0%    3.040µ ±  0%        ~ (p=0.109 n=10)
Coloring/exams/SanSegundo-8                                         7.403µ ±   1%    7.408µ ±  0%        ~ (p=0.324 n=10)
Coloring/exams/WelshPowell-8                                        1.823µ ±   0%    1.823µ ±  0%        ~ (p=0.867 n=10)
Coloring/exam_scheduler_1/Dsatur-8                                  17.83µ ±   0%    17.84µ ±  0%        ~ (p=0.128 n=10)
Coloring/exam_scheduler_1/DsaturExact-8                             98.82µ ±   0%    98.68µ ±  1%        ~ (p=0.315 n=10)
Coloring/exam_scheduler_1/Randomized-8                              4.232µ ±   0%    4.237µ ±  0%        ~ (p=0.065 n=10)
Coloring/exam_scheduler_1/RecursiveLargestFirst-8                   11.35µ ±   0%    11.33µ ±  0%   -0.17% (p=0.022 n=10)
Coloring/exam_scheduler_1/SanSegundo-8                              27.35µ ±   0%    27.32µ ±  1%        ~ (p=0.494 n=10)
Coloring/exam_scheduler_1/WelshPowell-8                             4.717µ ±   0%    4.716µ ±  0%        ~ (p=0.867 n=10)
Coloring/exam_scheduler_2/Dsatur-8                                  8.313µ ±   0%    8.315µ ±  0%        ~ (p=0.867 n=10)
Coloring/exam_scheduler_2/DsaturExact-8                             47.73µ ±   0%    47.75µ ±  0%        ~ (p=0.971 n=10)
Coloring/exam_scheduler_2/Randomized-8                              2.324µ ±   0%    2.326µ ±  0%        ~ (p=0.118 n=10)
Coloring/exam_scheduler_2/RecursiveLargestFirst-8                   5.167µ ±   0%    5.168µ ±  1%        ~ (p=0.288 n=10)
Coloring/exam_scheduler_2/SanSegundo-8                              12.70µ ±   0%    12.70µ ±  0%        ~ (p=0.644 n=10)
Coloring/exam_scheduler_2/WelshPowell-8                             2.729µ ±   0%    2.729µ ±  0%        ~ (p=0.267 n=10)
Coloring/Brélaz/Dsatur-8                                            12.02µ ±   0%    12.04µ ±  0%   +0.12% (p=0.025 n=10)
Coloring/Brélaz/DsaturExact-8                                       59.18µ ±   0%    59.18µ ±  0%        ~ (p=0.971 n=10)
Coloring/Brélaz/Randomized-8                                        3.052µ ±   0%    3.062µ ±  0%   +0.33% (p=0.000 n=10)
Coloring/Brélaz/RecursiveLargestFirst-8                             7.577µ ±   0%    7.552µ ±  0%   -0.34% (p=0.002 n=10)
Coloring/Brélaz/SanSegundo-8                                        18.30µ ±   0%    18.29µ ±  0%        ~ (p=0.870 n=10)
Coloring/Brélaz/WelshPowell-8                                       3.699µ ±   0%    3.695µ ±  0%        ~ (p=0.054 n=10)
Coloring/Lima_Carmo/Dsatur-8                                        7.529µ ±   2%    7.520µ ±  0%        ~ (p=0.392 n=10)
Coloring/Lima_Carmo/DsaturExact-8                                   32.60µ ±   0%    32.52µ ±  0%   -0.25% (p=0.024 n=10)
Coloring/Lima_Carmo/Randomized-8                                    2.076µ ±   0%    2.071µ ±  0%   -0.26% (p=0.027 n=10)
Coloring/Lima_Carmo/RecursiveLargestFirst-8                         4.184µ ±   0%    4.162µ ±  0%   -0.54% (p=0.000 n=10)
Coloring/Lima_Carmo/SanSegundo-8                                    11.18µ ±   0%    11.16µ ±  0%   -0.20% (p=0.002 n=10)
Coloring/Lima_Carmo/WelshPowell-8                                   2.476µ ±   0%    2.470µ ±  1%   -0.26% (p=0.033 n=10)
Coloring/San_Segundo/Dsatur-8                                       12.73µ ±   0%    12.70µ ±  0%   -0.27% (p=0.002 n=10)
Coloring/San_Segundo/DsaturExact-8                                  119.2µ ±   0%    118.9µ ±  0%   -0.24% (p=0.019 n=10)
Coloring/San_Segundo/Randomized-8                                   3.417µ ±   0%    3.415µ ±  1%        ~ (p=0.725 n=10)
Coloring/San_Segundo/RecursiveLargestFirst-8                        9.219µ ±   0%    9.184µ ±  0%   -0.39% (p=0.008 n=10)
Coloring/San_Segundo/SanSegundo-8                                   22.45µ ±   0%    22.40µ ±  0%        ~ (p=0.172 n=10)
Coloring/San_Segundo/WelshPowell-8                                  3.588µ ±   0%    3.587µ ±  0%        ~ (p=0.752 n=10)
Coloring/tetrahedron/Dsatur-8                                       2.373µ ±   1%    2.372µ ±  0%        ~ (p=0.360 n=10)
Coloring/tetrahedron/DsaturExact-8                                  13.52µ ±   0%    13.49µ ±  0%   -0.25% (p=0.000 n=10)
Coloring/tetrahedron/Randomized-8                                   927.4n ±   0%    926.2n ±  0%        ~ (p=0.343 n=10)
Coloring/tetrahedron/RecursiveLargestFirst-8                        1.853µ ±   0%    1.845µ ±  0%   -0.43% (p=0.000 n=10)
Coloring/tetrahedron/SanSegundo-8                                   3.312µ ±   1%    3.309µ ±  0%        ~ (p=0.469 n=10)
Coloring/tetrahedron/WelshPowell-8                                  1.064µ ±   0%    1.060µ ±  0%   -0.33% (p=0.001 n=10)
Coloring/triangle/Dsatur-8                                          1.473µ ±   0%    1.475µ ±  0%        ~ (p=0.116 n=10)
Coloring/triangle/DsaturExact-8                                     7.213µ ±   0%    7.241µ ±  1%   +0.40% (p=0.000 n=10)
Coloring/triangle/Randomized-8                                      636.5n ±   0%    635.8n ±  0%        ~ (p=0.109 n=10)
Coloring/triangle/RecursiveLargestFirst-8                           1.152µ ±   1%    1.147µ ±  0%   -0.48% (p=0.000 n=10)
Coloring/triangle/SanSegundo-8                                      1.813µ ±   0%    1.815µ ±  0%        ~ (p=0.665 n=10)
Coloring/triangle/WelshPowell-8                                     769.1n ±   0%    767.8n ±  1%        ~ (p=0.085 n=10)
Coloring/square/Dsatur-8                                            2.190µ ±   0%    2.191µ ±  0%        ~ (p=0.322 n=10)
Coloring/square/DsaturExact-8                                       10.79µ ±   1%    10.77µ ±  0%        ~ (p=0.165 n=10)
Coloring/square/Randomized-8                                        856.4n ±   0%    852.8n ±  0%   -0.42% (p=0.001 n=10)
Coloring/square/RecursiveLargestFirst-8                             1.438µ ±   0%    1.415µ ±  0%   -1.63% (p=0.001 n=10)
Coloring/square/SanSegundo-8                                        2.808µ ±   0%    2.797µ ±  0%   -0.41% (p=0.000 n=10)
Coloring/square/WelshPowell-8                                       1.020µ ±   0%    1.016µ ±  0%   -0.39% (p=0.000 n=10)
Coloring/cycle-5/Dsatur-8                                           3.222µ ±   1%    3.209µ ±  0%   -0.39% (p=0.000 n=10)
Coloring/cycle-5/DsaturExact-8                                      19.10µ ±   0%    19.03µ ±  0%   -0.38% (p=0.002 n=10)
Coloring/cycle-5/Randomized-8                                       1.127µ ±   0%    1.123µ ±  0%   -0.35% (p=0.003 n=10)
Coloring/cycle-5/RecursiveLargestFirst-8                            2.119µ ±   0%    2.109µ ±  0%   -0.45% (p=0.000 n=10)
Coloring/cycle-5/SanSegundo-8                                       4.142µ ±   0%    4.134µ ±  0%   -0.18% (p=0.005 n=10)
Coloring/cycle-5/WelshPowell-8                                      1.359µ ±   0%    1.358µ ±  0%   -0.07% (p=0.019 n=10)
Coloring/cycle-6/Dsatur-8                                           4.303µ ±   0%    4.290µ ±  0%   -0.29% (p=0.002 n=10)
Coloring/cycle-6/DsaturExact-8                                      18.22µ ±   0%    18.26µ ±  1%   +0.20% (p=0.003 n=10)
Coloring/cycle-6/Randomized-8                                       1.376µ ±   0%    1.371µ ±  0%   -0.36% (p=0.000 n=10)
Coloring/cycle-6/RecursiveLargestFirst-8                            2.490µ ±   0%    2.481µ ±  0%   -0.38% (p=0.000 n=10)
Coloring/cycle-6/SanSegundo-8                                       5.572µ ±   0%    5.550µ ±  0%   -0.40% (p=0.000 n=10)
Coloring/cycle-6/WelshPowell-8                                      1.665µ ±   0%    1.660µ ±  0%   -0.33% (p=0.001 n=10)
Coloring/wheel-5/Dsatur-8                                           3.446µ ±   0%    3.432µ ±  0%   -0.41% (p=0.003 n=10)
Coloring/wheel-5/DsaturExact-8                                      22.82µ ±   0%    22.72µ ±  0%   -0.44% (p=0.009 n=10)
Coloring/wheel-5/Randomized-8                                       1.227µ ±   1%    1.221µ ±  0%   -0.49% (p=0.000 n=10)
Coloring/wheel-5/RecursiveLargestFirst-8                            2.395µ ±   0%    2.379µ ±  0%   -0.65% (p=0.000 n=10)
Coloring/wheel-5/SanSegundo-8                                       5.155µ ±   0%    5.155µ ±  0%        ~ (p=0.617 n=10)
Coloring/wheel-5/WelshPowell-8                                      1.391µ ±   0%    1.391µ ±  0%        ~ (p=0.869 n=10)
Coloring/wheel-6/Dsatur-8                                           4.816µ ±   0%    4.824µ ±  0%   +0.16% (p=0.004 n=10)
Coloring/wheel-6/DsaturExact-8                                      37.64µ ±   0%    37.65µ ±  0%        ~ (p=0.926 n=10)
Coloring/wheel-6/Randomized-8                                       1.586µ ±   0%    1.592µ ±  0%   +0.35% (p=0.000 n=10)
Coloring/wheel-6/RecursiveLargestFirst-8                            3.311µ ±   0%    3.313µ ±  0%        ~ (p=0.208 n=10)
Coloring/wheel-6/SanSegundo-8                                       7.499µ ±   0%    7.521µ ±  0%   +0.29% (p=0.001 n=10)
Coloring/wheel-6/WelshPowell-8                                      1.800µ ±   0%    1.801µ ±  0%        ~ (p=0.526 n=10)
Coloring/wheel-7/Dsatur-8                                           6.295µ ±   0%    6.298µ ±  0%        ~ (p=0.381 n=10)
Coloring/wheel-7/DsaturExact-8                                      39.04µ ±   0%    39.11µ ±  0%   +0.16% (p=0.002 n=10)
Coloring/wheel-7/Randomized-8                                       1.930µ ±   0%    1.941µ ±  0%   +0.54% (p=0.000 n=10)
Coloring/wheel-7/RecursiveLargestFirst-8                            3.937µ ±   0%    3.946µ ±  0%   +0.24% (p=0.005 n=10)
Coloring/wheel-7/SanSegundo-8                                       10.10µ ±   1%    10.14µ ±  0%   +0.37% (p=0.009 n=10)
Coloring/wheel-7/WelshPowell-8                                      2.169µ ±   0%    2.171µ ±  0%        ~ (p=0.089 n=10)
Coloring/sudoku_board/Dsatur-8                                      21.54m ±   0%    21.61m ±  0%   +0.34% (p=0.043 n=10)
Coloring/sudoku_board/DsaturExact-8                                 86.12m ±   1%    86.51m ±  0%        ~ (p=0.280 n=10)
Coloring/sudoku_board/Randomized-8                                  585.7µ ±   0%    584.9µ ±  0%        ~ (p=0.165 n=10)
Coloring/sudoku_board/RecursiveLargestFirst-8                       3.907m ±   0%    3.874m ±  0%   -0.86% (p=0.000 n=10)
Coloring/sudoku_board/SanSegundo-8                                  25.67m ±   0%    25.72m ±  0%   +0.21% (p=0.015 n=10)
Coloring/sudoku_board/WelshPowell-8                                 691.3µ ±   0%    694.1µ ±  0%   +0.40% (p=0.000 n=10)
Coloring/sudoku_problem/Dsatur-8                                    32.68m ±   0%    32.94m ±  0%   +0.80% (p=0.000 n=10)
Coloring/sudoku_problem/DsaturExact-8                                6.988 ±  45%     6.543 ± 65%        ~ (p=0.631 n=10)
Coloring/sudoku_problem/Randomized-8                                833.7µ ±   0%    831.1µ ±  0%   -0.31% (p=0.011 n=10)
Coloring/sudoku_problem/RecursiveLargestFirst-8                     5.831m ±   0%    5.750m ±  0%   -1.39% (p=0.000 n=10)
Coloring/sudoku_problem/SanSegundo-8                                35.74m ±   0%    35.99m ±  0%   +0.71% (p=0.000 n=10)
Coloring/sudoku_problem/WelshPowell-8                               971.7µ ±   1%    969.7µ ±  0%        ~ (p=0.143 n=10)
Coloring/cs_shc/Dsatur-8                                            3.343µ ±   0%    3.348µ ±  0%   +0.15% (p=0.044 n=10)
Coloring/cs_shc/DsaturExact-8                                       20.27µ ±   0%    20.28µ ±  0%        ~ (p=0.108 n=10)
…